-Small refactor and Unity UI

+++ Assets/Codes/_main/MinityScene.cs
@@ -40,8 +40,8 @@ namespace MinityEngine
         public static extern void SetTime(float time, float deltaTime);
 
         [DllImport(MinityPluginName)]
-        public static extern void SetScreenValues(float width, float height, float aspect);
-
+        public static extern void SetScreenValues(int width, int height, float aspect);
+         
         [SerializeField] private Material _mat;
 
         //[DllImport(MinityPluginName)]
